feat: update `CClassicHealthBar` class with new design and better indicating reloading/planting status
feat: add bomb and timer icons to topbar

feat: add possibility to adjust font weight (but haven't added command)

feat: display skull sprite instead of HP in case of death

fix: HP number aligned to center

fix: use custom topbar background to solve text z-index issue

refactor: remove unused `CCSPHealthBar`

+++ src/textures.cpp
@@ -16,7 +16,7 @@ void DrawTexture(hud_texture_s texture, int x, int y, float alpha, int flags) {
 
     glBindTexture(GL_TEXTURE_2D, texture.texID);
 
-    glColor4f(1.0, 1.0, 1.0, static_cast<GLfloat>(alpha));
+    glColor4f(1.0, 1.0, 1.0, alpha);
 
     glBegin(GL_QUADS);
     glTexCoord2f(0, 0.F);
